backend/requirements.txt
# "This file is used to specify the dependencies required by the project to run properly,
# for example, dependencies
# to update this run 'pip3 freeze >> requirements.txt'
# in the same directory as this file.

# the command I used to run the server is flask --app run --debug

mysql-connector-python~=8.3.0
pip~=23.2.1